+++ shared/js/ui/react-elements/modal/Definitions.ts
@@ -23,6 +23,7 @@ import {ModalAvatarUploadEvents, ModalAvatarUploadVariables} from "tc-shared/ui/
 import {ModalInputProcessorEvents, ModalInputProcessorVariables} from "tc-shared/ui/modal/input-processor/Definitios";
 import {ModalServerInfoEvents, ModalServerInfoVariables} from "tc-shared/ui/modal/server-info/Definitions";
 import {ModalAboutVariables} from "tc-shared/ui/modal/about/Definitions";
+import {ModalServerBandwidthEvents} from "tc-shared/ui/modal/server-bandwidth/Definitions";
 
 export type ModalType = "error" | "warning" | "info" | "none";
 export type ModalRenderType = "page" | "dialog";
@@ -225,5 +226,8 @@ export interface ModalConstructorArguments {
     "modal-server-info": [
         /* events */ IpcRegistryDescription<ModalServerInfoEvents>,
         /* variables */ IpcVariableDescriptor<ModalServerInfoVariables>
+    ],
+    "modal-server-bandwidth": [
+        /* events */ IpcRegistryDescription<ModalServerBandwidthEvents>
     ]
 }